requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
greenlet==3.1.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
